feat(late-interaction): default storage to float32 for query speed (#25)
Prefer the fastest, bit-exact query path by default; float16 / int8 stay opt-in
for a smaller footprint. New indexes default to float32 (~1.5ms/query here vs
~3.6ms for float16), and the choice still persists with the index. Updated tests,
benchmark default + result, and docs.

+++ benchmarks/late_interaction/run.py
@@ -287,8 +287,8 @@ def main() -> None:
     parser.add_argument("--k", type=int, default=10)
     parser.add_argument(
         "--storage",
-        choices=("float16", "float32", "int8"),
-        default="float16",
+        choices=("float32", "float16", "int8"),
+        default="float32",
         help="patch-matrix storage precision",
     )
     parser.add_argument("--seed", type=int, default=0)
